Added settingsView.js
